+++ mle/model/ollama.py
@@ -1,0 +1,55 @@
+import importlib.util
+
+from .common import Model
+
+
+class OllamaModel(Model):
+    def __init__(self, model, host_url=None):
+        """
+        Initialize the Ollama model.
+        Args:
+            host_url (str): The Ollama Host url.
+            model (str): The model version.
+        """
+        super().__init__()
+
+        dependency = "ollama"
+        spec = importlib.util.find_spec(dependency)
+        if spec is not None:
+            self.model = model if model else 'llama3'
+            self.model_type = 'Ollama'
+            self.ollama = importlib.import_module(dependency)
+            self.client = self.ollama.Client(host=host_url)
+        else:
+            raise ImportError(
+                "It seems you didn't install ollama. In order to enable the Ollama client related features, "
+                "please make sure ollama Python package has been installed. "
+                "More information, please refer to: https://github.com/ollama/ollama-python"
+            )
+
+    def query(self, chat_history, **kwargs):
+        """
+        Query the LLM model.
+        Args:
+            chat_history: The context (chat history).
+        """
+
+        # Check if 'response_format' exists in kwargs
+        format = None
+        if 'response_format' in kwargs and kwargs['response_format'].get('type') == 'json_object':
+            format = 'json'
+
+        return self.client.chat(model=self.model, messages=chat_history, format=format)['message']['content']
+
+    def stream(self, chat_history, **kwargs):
+        """
+        Stream the output from the LLM model.
+        Args:
+            chat_history: The context (chat history).
+        """
+        for chunk in self.client.chat(
+                model=self.model,
+                messages=chat_history,
+                stream=True
+        ):
+            yield chunk['message']['content']
